refactor: move ChartPlaceholder out from MemberView to src/
